refactor: 이미지 설정 관련 코드 분리 - 이미지 업로드와 이미지 목록을 하나의 상위 컴포넌트를 통해 렌더링하도록 edit 페이지의 메인 컴포넌트에서 분리
